Bug 1577051 - Fix an implicit-conversion-changes-value warning with new-enough clang by explicitly converting.  r=froydnj
Differential Revision: https://phabricator.services.mozilla.com/D43701

UltraBlame original commit: d72f6bb552f38a54c6df3b1d48e28849da0ce189

+++ mfbt/FastBernoulliTrial.h
@@ -363,9 +363,11 @@ class FastBernoulliTrial {
 
 
 
+
+
     double skipCount =
         std::floor(std::log(mGenerator.nextDouble()) * mInvLogNotProbability);
-    if (skipCount < SIZE_MAX)
+    if (skipCount < double(SIZE_MAX))
       mSkipCount = skipCount;
     else
       mSkipCount = SIZE_MAX;
